TCVP-1680 LookupService using ORDS OpenAPI

+++ src/backend/oracle-data-api/src/test/java/ca/bc/gov/open/jag/tco/oracledataapi/BaseTestSuite.java
@@ -4,6 +4,7 @@ import javax.transaction.Transactional;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 
@@ -15,16 +16,22 @@ import ca.bc.gov.open.jag.tco.oracledataapi.repository.JJDisputeRepository;
 @Transactional
 public class BaseTestSuite {
 
+	@Value("${ords.enabled}")
+    protected boolean ordsEnabled;
+
 	@Autowired
 	protected DisputeRepository disputeRepository;
-	
+
 	@Autowired
 	protected JJDisputeRepository jjDisputeRepository;
 
     @BeforeEach
     protected void beforeEach() throws Exception {
-    	disputeRepository.deleteAll();
-    	jjDisputeRepository.deleteAll();
+    	// only delete the repo if this is a local H2 repository.
+    	if (!ordsEnabled) {
+	    	disputeRepository.deleteAll();
+	    	jjDisputeRepository.deleteAll();
+    	}
     }
 
 }
